refactor: Refactoring logic code
