+++ src/hooks/logic/useAvatarColor.js
